moved constants from Pin interface to Tr03110 interface, removed Pin interface

+++ de.persosim.simulator/src/de/persosim/simulator/protocols/Tr03110.java
@@ -13,7 +13,7 @@ public interface Tr03110 {
 	public final static byte[] id_BSI                              = {0x04, 0x00, 0x7F, 0x00, 0x07};
 	public final static byte[] id_PK                               = Utils.appendBytes(id_BSI, new byte[]{0x02, 0x02, 0x01});
 	
-	public final static int ID_PIN = 3;
-	public final static int ID_CAN = 2;
+	public final static byte ID_PIN = 3;
+	public final static byte ID_CAN = 2;
 
 }
